Correct edge case that could result in NaN output
Colors with NaN rgb values were rarely returned from hswlToRgb. Float
and other rounding could cause the luminance error adjustment to
be negative in certain instances, which when fed back in to solveHswl
would take the root of the negative value. Also clean up some
sloppines with the the variable declarations.